Handle logic for showing and creating schedules
Once a user logs in, there needs to be some logic to determine if they
need to create a new schedule or show the exisiting one. If there is no
schedule related to the user, they should create one. This is also true
if their last completed schedule is was completed more than two days
ago.

Otherwise the schedule is still current and so should be displayed.
However, if it has been completed, but hasn't finsihed it's final break period, then
a screen informing the date of the next test should be shown.